update: FirebaseHelperを追加してfirebaseConfigureをAppDelegateから書き出し
