+++ QuickGourmet/AppDelegate.swift
@@ -5,7 +5,6 @@
 //  Created by sasaki.ken on 2021/12/22.
 //
 
-import Firebase
 import NendAd
 import SwiftUI
 
@@ -22,14 +21,16 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil) -> Bool {
-        NADInterstitial.sharedInstance().loadAd(withSpotID: NEND_INTERSTITIAL_STILLNESS_SPOTID, apiKey: NEND_INTERSTITIAL_STILLNESS_AD_APIKEY)
-
-        // 最新
-        firebaseConfigure()
+        FirebaseHelper.firebaseConfigure()
+        setUpNendAd()
         recordLaunchCount()
         saveDeviceId()
 
         return true
+    }
+
+    private func setUpNendAd() {
+        NADInterstitial.sharedInstance().loadAd(withSpotID: NEND_INTERSTITIAL_STILLNESS_SPOTID, apiKey: NEND_INTERSTITIAL_STILLNESS_AD_APIKEY)
     }
 
     private func recordLaunchCount() {
@@ -42,23 +43,5 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                 userRepository.deviceId = deviceId
             }
         }
-    }
-
-    private func firebaseConfigure() {
-        #if DEBUG
-            let filePath = Bundle.main.path(forResource: "GoogleService-Stage-Info", ofType: "plist")
-        #else
-            let filePath = Bundle.main.path(forResource: "GoogleService-Info", ofType: "plist")
-        #endif
-
-        guard let filePath = filePath else {
-            return
-        }
-
-        guard let options = FirebaseOptions(contentsOfFile: filePath) else {
-            return
-        }
-
-        FirebaseApp.configure(options: options)
     }
 }
